Skip files with EXIF data
Converter now checks and skips over files which already have an EXIF
CreateDate ("Date Taken") attached to the file. This fixes #4.

An issue was also corrected where a divide-by-zero error could occur if
all files in the directory needed to be skipped.

Changed "skip" message to be more generic and vague. This will need to
be improved later on by adding either a log or some other method of the
user to find out more details.

Also increased version number.